The previous epsilon value doesn't really work very well for finding the nearest vertex.
Added a 'delete all shapes' function which can be invoked using ALT-D.

Copy boxes from previous image wasn't really very useful since it used the previous one in the list.  Amended so that we track the previously open file and copy the boxes from it instead.  This still isn't ideal, and the real solution is to enable multiple selections and use the clipboard to support copy & paste.

Modified the selected fill colour so it is easier to see the image underneath.